audio/requirements.txt
grpcio==1.35.0
grpcio-tools==1.35.0
protobuf==3.15.1
